backend/requirements.txt
azure-functions==1.20.0
jinja2==3.1.4
pydantic==2.7.1
